Refactor stuff out into storage.rs
